Stay on active menu item after page load
Added option for staying on the currently selected menu item when the page loads into that page. When opened again that menu level will be present. Only issue is with redirects. If the page is redirected to another URL, that URL will be set as the active menu item, not the page that was selected in the menu.

Signed-off-by: Blake Cerecero <Blake@DigitalBlake.com>