Prepend close button in JS
For site where the original button for the menu us not visible after opening the menu, I have added the option to prepend a close button to the UL for each menu section. The only thing I don't like about this is that for every left-right animation that occurs there is a new close menu button. I would rather this stay static and be the same button for every UL.

Signed-off-by: Blake Cerecero <Blake@DigitalBlake.com>