Minimized number of DOM searches
Went through and turned many of the DOM searches from the root of the menu into variables so that I only need to perform the search once then manipulate the DOM with the variable as needed. Reducing the number of times I must search the entire DOM for an element.

Signed-off-by: Blake Cerecero <Blake@DigitalBlake.com>