Fixed Mobile Opera Issue and added new Icons
The icons are SVG with a PNG fallback. For the Opera Mobile issue it turns out that I just need to use translateX rather than translateZ and the issue would not have occured. This is for the css transform property. I also modified that JS fallback for IE to use an if else statment looking to the open class on the menu to open or close on click.

Signed-off-by: Blake Cerecero <Blake@DigitalBlake.com>